removing the shiftSpace eventListener for invoking editor with modifier key
I do not need this or plan to use it right now so removing to clean things up but this commit will show me what the code was if I ever want to pursue it in the future. I actually don't even know if this works as it is old code I haven't retested.